refactor: Refactor import statements in test_result.py to use core module
- Updated import paths for COMPONENTS and DPFResult from
  ansys.mapdl.core.reader.result to ansys.mapdl.core.reader.core.
- This change aligns with the new module structure and ensures
  consistency across the test suite.
