+++ src/ansys/mapdl/core/reader/__init__.py
@@ -20,4 +20,4 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
-from .result import DPFResult
+from ansys.mapdl.core.reader.core import DPFResult
