+++ src/ansys/mapdl/core/reader/core.py
